\name{HUR}
\alias{HUR}
\docType{data}
\title{
Hidden Unemployment in Sao Paulo
}
\description{
Rate of hidden unemployment due to substandard work conditions in Sao Paulo, Brazil (Rocha and Cribari-Neto, 2009). 
}
\usage{data(HUR)}
\source{Institute of Applied Economic Research (Ipea), Brazil. Data obtained from the IPEAdata website \url{http://www.ipeadata.gov.br}.
}
\references{
Rocha, A.V. and Cribari-Neto, F. (2009). Beta autoregressive moving average models. \emph{Test} \bold{18}, 529--545.
}