\name{ind.cormat}
\alias{ind.cormat}
\title{
Gaussian Copula Regression with Working independence Correlation Matrix
}
\description{
Sets working independence correlation matrix in Gaussian copula marginal regression models.
}
\usage{
ind.cormat()
}
\value{
An object of class \code{\link{cormat.gcmr}} representing an identity correlation matrix.
}
\references{
Masarotto, G. and Varin, C. (2011). Gaussian copula marginal regression. Preprint.
}
\author{
Guido Masarotto and Cristiano Varin.
}
\seealso{
\code{\link{gcmr}}
}
\keyword{regression}
\keyword{nonlinear}