\name{gcmr-package}
\alias{gcmr-package}
\docType{package}
\title{
Gaussian Copula Marginal Regression
}
\description{
 Fits Gaussian copula marginal regression models described by Masarotto and Varin (2011). 
}
\details{
This package contains \R functions related to the paper Masarotto and Varin (2011). The main function is \code{\link{gcmr}}. This allows for likelihood inference in the class of Gaussian copula marginal regression models for non-normal responses with a flexible specification of the dependence structure.   
}
\author{
Guido Masarotto and Cristiano Varin.
}
\references{
Masarotto, G. and Varin, C. (2011). Gaussian copula marginal regression. Preprint. 
}
\keyword{package}
