\name{ind.cormat}
\alias{ind.cormat}
\title{
Working Independence Correlation
}
\description{
Sets working independence correlation in Gaussian copula marginal regression models.
}
\usage{
ind.cormat()
}
\value{
An object of class \code{\link{cormat.gcmr}} representing an identity correlation matrix.
}
\references{
Masarotto, G. and Varin, C. (2012). Gaussian copula marginal regression. \emph{Electronic Journal of Statistics} \bold{6}, 1517--1549. \url{http://projecteuclid.org/euclid.ejs/1346421603}.

Masarotto, G. and Varin C. (2017). Gaussian Copula Regression in R. \emph{Journal of Statistical Software}, \bold{77}(8), 1--26. \doi{10.18637/jss.v077.i08}.
}
\author{
Guido Masarotto and Cristiano Varin.
}
\seealso{
\code{\link{gcmr}}.
}
\keyword{regression}
\keyword{nonlinear}